refactor: 안쓰는 모달 컴포넌트 스타일, 파일 삭제 - BasicDialog.tsx는 BasicModal.tsx로 대체되어 삭제 - BasicModal.style.tsx 파일 삭제 - BasicModal.style.ts 파일 삭제 - Remove unused code
