+++ .gitignore
@@ -24,3 +24,4 @@ dist-ssr
 *.sw?
 
 .env
+premissions.json
